gtf {IntAssoPlot}	R Documentation

Genome annotation file.

Description

A dataset containing the annotation file WITHOUT the column name. The example file is Maize genome annotation, and the version is AGPv2.IMPORTANT, the annotation file could be read with header=FALSE. The variables are as follows:

Usage

gtf

Format

:

the first column

The name of the sequence. Commonly, this is the chromosome ID or contig ID

the second column

The source column should be a unique label indicating whether the annotation is protein-coding or something else.

the third column

The following feature types are required: 'CDS', 'start_codon', 'stop_codon'. The features '5UTR', '3UTR', 'inter', 'inter_CNS', 'intron_CNS' and 'exon' are optional.

the fourth column

Integer start coordinates of the feature relative to the beginning of the sequence named

the fifth co0umn

Integer end coordinates of the feature relative to the beginning of the sequence named

the sixth column

The score field indicates a degree of confidence in the feature's existence and coordinates.

the seventh column

The strand indicates the annotation is located on forward or reverse strand.

the eigth column

frame

the ninth column

attributes

[Package IntAssoPlot version 0.99.21 Index]